backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
twilio==8.10.0